[Enhancement]: Extracted sliver list into new widget.
