+++ lib/src/widgets/restaurant_list.dart
@@ -1,0 +1,89 @@
+import 'package:flutter/material.dart';
+import 'package:you_choose/src/models/restaurant.dart';
+
+class RestaurantList extends StatefulWidget {
+  final List<Restaurant> restaurantsList;
+
+  const RestaurantList({Key? key, required this.restaurantsList})
+      : super(key: key);
+
+  @override
+  State<RestaurantList> createState() => _RestaurantListState();
+}
+
+class _RestaurantListState extends State<RestaurantList> {
+  Widget _buildListItem(BuildContext context, Restaurant document) {
+    final String? name = document.name;
+    final int? price = document.price;
+    final List<String>? tags = document.tags;
+    final String? description = document.description;
+
+    return Container(
+        height: 200,
+        margin: const EdgeInsets.only(left: 20, right: 20, bottom: 15),
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(30),
+          color: Colors.white,
+        ),
+        child: Container(
+          padding: const EdgeInsets.only(top: 30, left: 20, right: 20),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: <Widget>[
+              Text(
+                name!,
+                maxLines: 1,
+                overflow: TextOverflow.ellipsis,
+                style: const TextStyle(
+                    color: Color(0xFF332d2b),
+                    fontWeight: FontWeight.w700,
+                    fontSize: 20),
+              ),
+              const SizedBox(height: 2),
+              Wrap(
+                children: List.generate(
+                    price!,
+                    (index) => const Icon(
+                          Icons.currency_pound,
+                          color: Colors.blueAccent,
+                          size: 15,
+                        )),
+              ),
+              const SizedBox(height: 10),
+              Text(
+                description!,
+                maxLines: 2,
+                overflow: TextOverflow.ellipsis,
+                style:
+                    const TextStyle(fontWeight: FontWeight.w500, fontSize: 12),
+              ),
+              const SizedBox(height: 10),
+              Wrap(
+                  spacing: 8.0,
+                  runSpacing: 4.0,
+                  children: List.generate(tags!.length, (index) {
+                    return Chip(
+                      label: Text(tags[index]),
+                      backgroundColor: Colors.lightBlueAccent,
+                      elevation: 2.0,
+                      labelStyle: const TextStyle(
+                        fontSize: 12,
+                        fontWeight: FontWeight.w400,
+                      ),
+                    );
+                  }))
+            ],
+          ),
+        ));
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return SliverList(
+      delegate: SliverChildBuilderDelegate(
+          (BuildContext context, int index) =>
+              _buildListItem(context, widget.restaurantsList[index]),
+          childCount: widget.restaurantsList.length),
+    );
+  }
+}
